Refactor PostCalendar smarty to twigs
This is the first of several commits that will refactor the postcalendar
smarty templates into twig templates.  I started with the day template
first and got 75% of the code into the twig template.  The core event
loop had to be stuffed into a twig function for now until we can pull
out the loop iteration into the business logic.

Currently the day print functionality is broken until we refactor that
logic.  I also need to do the week and monthly view.