Fixes #57, #58 submit changes button
Made it so the submit changes button always shows but it is disabled
until data has changed.

Made it so the data change watch on the submit changes button is based
on the input event instead of the change event.  This makes it so we can
capture input changes on mobile devices and enable the save button
without having the mobile user have to tap somewhere else to enable the
button.

Fixes #57
Fixes #58