WIP for snomed code speedups
Added a method to speedup searching for snomed codes.  It was done with
help of claude.ai so will need to verify edge cases are handled
properly.  Its possible we could consolidate the two functions but where
snomed code sets are such a large database we need to have really fast
lookups and this provided magnitude faster searching than the previous
generic query.