Improved code lookup for snomeds and large tables
Changed up the search mechanism for code types with external tables that
define a fulltext search on the search term.  Use the database search
features for improved lookup performance.

Also made it so tables can define skipping the table count properties in
the find code files (find_code_dynamic and select_codes).

Not sure I like it, but for now ICD10 and other tables show a list of
codes when the finder first pops up, whereas snomed when it first shows
will show an empty table.  You need to put in a search term.

I brought back the auto-complete typing though I bumped it up to 350ms
delay instead of 200ms delay.  This fixes the enter key closing the
dialog bug as well as makes the finder more usable.  With the improved
search performance the autocomplete now can function well.  Will have to
test this when there's a lot of concurrency.

Overall its a big improvement.  More testing will have to see how it
works.