Experiment 1 for api code coverage
Implemented code coverage collection for the api test suite.  Need to
figure out how to use the phpunit cache warmup for grabbing all of the
src directory.  When the src directory is included the tests slow down
to a crawl and we get api timeouts.  For now in this experiment just did
an initial run for building the coverage files which get stored in the
coverage directory. I hard-code the files in the
CoverageHelper::createTargetedCodeCoverage() method and will need to
figure out the directory traversal later.

Afterwards to compile all the coverage into an aggegrated report the
following commands can be run:
```
 php -d memory_limit=512M vendor/bin/phpcov merge coverage --html=coverage/html --clover=coverage/clover.xml && rm cov
erage/*.cov
```

I add two environment variables for the docker-compose docker
environment.  One to turn on the coverage report collection which will
be turned off by default (and respected in the CoverageHelper class) and
another to tell xdebug to collect coverage metrics.