Initial coverage support.
Added coverage support to the dispatch.php file that gets called on
every inferno request.  It only runs if the environment variable is set
so we can collect coverage information.

The coverage process is extremely slow taking 30 seconds per http
request from inferno which means it will take around ~2 hours for the
coverage report to run for inferno.  Supposedly the pcov extension will
take a lot less time so going to see if I can experiment with that and
see how that works.