Smoke screen test fix execution errors.
Wrote up a ApiApplication class that does all of the api bootstrapping
and setup enabling us to test it with a smoke screen test.

This identified a number of errors and so I was able to fix it.

had to modify the .htaccess to work like the symfony .htaccess instead
of passing the path as a URL argument, we treat it like a path which
makes the URI parsing from symfony work properly.

Unit tests have to be currently executed like this:
```
./vendor/bin/phpunit -c phpunit.api.xml --testsuite api --filter ApiApplicationTest
```

Since the application bootstraps the entire globals.php, we can't use
the standard phpunit config which starts the globals.php as part of
bootstrap otherwise we end up with double session problems.