Fix forced exit and errors/warnings
Fixed the site path so skipped locations are still skipped.  Would still
prefer to have just an overall single firewall setup but for now we will
leave this as is and hopefully we can replace it in a few months.

Telemetry was force exiting so wasn't seeing my phpunit results. Fixed
that also