Cleaning up restconfig, fix bearer token
Got the bearer token to work for standard api request on facility but
discovered that the RestConfig is making it an absolute chore to
refactor the apis due to how it isn't PSR compliant includes a bunch of
files and changes some of the global values.

I'm going to commit this chunk and see if I can refactor the
authorize.php endpoint to use the same AppApplication and handle the
authentication pieces via the AuthorizationController without needing to
depend on the RestConfig.

I also realized that we're pretty much duplicating a ton of work in the
ScopeRepository by checking the scopes against the _rest_routes.  It
keeps things in sync but its a ton of extra processing pile that is done
on every single request.

Instead if we just force people to put in scopes in the ScopeRepository
at the same time they add an endpoint we can eliminate a ton of
processing on every request which should speed things up quite a bit
when getting access tokens, registering, etc.