Implemented ApiApplication for authorize.php
Put in the ApiApplication using the OEHttpKernel for the authorize.php
system.  Introduced a OAuth2AuthorizationListener that intercepts the
kernel request and provides all of the oauth2 handling that was being
done in AuthorizationController.

Currently we have to figure out how to handle the session & cookies in
AuthorizationController because they use SessionUtil a singleton, we
can't handle the testability components of the request/response that the
rest of the system relies on.

This is also creating issues with the oauth2 logout piece so we have one
test failing in the ApiTestClientTest class.

PSR style fixes in this as well.

Fixed a number of bugs in the ScopeRepository.  Also got all of the api
test suites working for Facility, Patient, and Practitioner.