Issues #8639,#8646,#8642 SMART on FHIR changes
Removed AuthorizationController dependency on
SmartConfigurationController

Refactored globals.php to use a container object that we can use for
test dependency injection, classes should now use the OEGlobalsBag
object for referencing global parameters / configuration.  This improves
testability and allows for items to be mocked/stubbed out much easier.
Fixes #8646.

Changed UserService::getAuthGroupForUser from static to instance method
so we can mock the behavior in testing.

Refactored OAuth2 grants to use SessionInterface so we can mock the
session for testing.

Changed up the ScopeRepository to handle both v1 and v2 SMART on FHIR
scopes for #8639.  We still need to handle the actual scope verification
piece which I'll be working on next once I've verified the SMART v1 and
us core 3.1.1 are unaffected with these changes.

Changed up CSRFUtils to work with the SessionInterface so we can use it
in testing.

Fixed up the HttpSessionFactory so we could keep it backwards compatible
with the existing session management pieces.

Added the OEGlobalsBag to the OEHttpKernel so we can grab the
configuration variables wherever we need, but still have it as a test
hub for mocking.

Removed the SystemConfigurationContext as I couldn't see it used
anywhere.

Added ServerConfig::getJsonWebKeySetUrl as an instance method so we can
use it in the SmartConfiguration.

Updated the SMART on FHIR Capability well-known configuration to handle
all of the v2 requirements.  Fixes #8642.

Fixed up a lot of the BearerTokenAuthorizationStrategy so we could make
it more testable.  This was originally refactored from the dispatch.php
class and can definitely have a lot more improvements as the class is
just doing too much (as evidenced by how much we have to mock
dependencies in the test classes).

Major refactoring of AuthorizationController so we can use
SessionInterface and the OEGlobalsBag to make the class testable and use
the data that comes from the HTTPKernel instead of global variable and
session state. Changed the class up to always return response objects
instead of exiting anywhere in the class, again so we can test and have
the class sit inside of the api request/response lifecycle.