API changes - response types,session fixes,cleanup
Fixed encounter rest controller to use request session.

Changed HttpRestRouteHandler->dispatch() to use the request's getMethod
instead of getRequestMethod().

Added a new trait SystemLoggerAwareTrait to be used in the
FhirServiceBase class so that all FHIR classes can use the same system
logger and support message suppression during the unit tests.  When we
intentionally trigger errors we don't want to clog up the test output
with expected error messages.

This also makes the system a bit faster as every systemlogger trigger's
its own shutdown handler via Monolog and having a single instance of the
logger via the requests makes things faster.

Cleaned up unused methods in RestConfig.

Changed the RestControllerHelper::responseHandler and
RestControllerHelper::handleFhirProcessingResult to return response
objects.  This was really the only way to handle returning the status
code pieces that has been built into the apis in the new api
architecture.  This is a breaking change that will impact module api
writers and will need to be communicated out to them.

Fixed up all the fhir rest controller test methods that expected the raw
object responses instead of the response bundles.