Fix htaccess issues on dispatch/authorize
The REWRITE_COMMAND in the htaccess rewrite is no longer used for the
api and authorize in order to leverage the symfony commands.

I'll have to consider the ramifications of this as it will require
people to update their nginx configurations for existing users and I'm
not sure if we want to break that for users.  However, it'd take some
heavier refactoring on the symfony request process if we want to go that
route.