Fix CapabilityStatement metadata
Removed the deprecated security extension for SMART on FHIR from the CapabilityStatement metadata.  Blows my mind how inferno is now failing with this but apparently there's a bug in the library inferno uses for its resource validation that is triggering a 500 error on the certification software.  The issue is documented here: https://github.com/onc-healthit/onc-certification-g10-test-kit/issues/655 where any deprecated extension is treated as a 500 error.  Now that the URIs are documented in the well-known file its safe to just remove the extension.

This now makes the 3.1.1 us core pass completely again with smart on
fhir 1.0.0.  Working on SMART on FHIR 2.1 now to make that work
properly.