OAuth2 remove superglobals, unit/integration tests
Added OEGlobalsBag to SMARTSessionTokenContextBuilder in order to remove
the superglobals that are used there.

Changed up the UserEntity class to have an injected fhirBaseURL so we
can populate the fhirUser claim properly and remove the superglobals
reference.

Changed up CustomRefereshTokenGrant to throw an exception if the context
can't be decoded.

Changed up IdTokenSMARTResponse to remove superglobals reference, built
out a unit test and integration test for the class to verify it is
functioning properly.

Deprecated IdentityRepository and added the corresponding methods to
UserRepository since they both instantiate UserEntity objects for
OIDC/OAuth2 purposes.

This should fix the failing unit tests due to the fact the
site_address_oauth property was not being set.