WIP US Core 8.0.0 fhir Condition resource.
This is part of issue #8787.

Changed up the FhirCondition logic to use the Mapped resource pattern
like we do with the Observation resources.  The condition is split up
into the encounter diganosis and problems/health concerns service
classes.  We may need to split problems/health concerns up even more
depending on where we pull the logic from for these resources.

Created a new Trait MappedServiceCategoryTrait for resources that only
use categories and not codes.

Implemented a ProfileMapper that I originally thought I was going to use
for adding / manipulating the data in the profile, but so far I haven't
had to do any of that so we may just remove the entire class.

Also wrote unit tests for the sub-resource classes for Condition.  Still
need more unit tests to handle additional code coverage.